slurm classes renaming for both parent and child class. Reverted to older slurm.py format for simplicity
